Fix selection update issue in KLinkItemSelectionModel.
This issue appear when using a model to link a proxy.
In some case (especially from QSFPM) a pair of
layoutAboutToBeChanged/layoutChanged is transformed into a pair of
layoutAboutToBeChanged/layoutChanged (empty model) and
rowsAboutToBeInserted/rowsInserted.
The proxy is so taking much longer time to reflect its changes making
the KLinkItemSelectionModel trying to update the proxy selection in
the middle of the layoutChanged/rowsInserted, exactly when it's empty
(involving unmapped selection which is then simply lost).

Task-Id: KDEND-59